requirements.txt
pandas
PyQt5
requests
openai
python-dotenv
pdfplumber
tiktoken
openpyxl
beautifulsoup4
lxml